+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/contract/model/exception/ContractAgreementException.java
@@ -1,9 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022,2024
- *       2022: ZF Friedrichshafen AG
- *       2022: ISTOS GmbH
- *       2022,2024: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *       2022,2023: BOSCH AG
+ * Copyright (c) 2022,2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
@@ -21,24 +17,17 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.edc.client.exceptions;
+package org.eclipse.tractusx.irs.edc.client.contract.model.exception;
 
-import lombok.Getter;
-import org.eclipse.edc.policy.model.Policy;
+import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
 
 /**
- * Usage Policy Exception errors in the contract negotiation.
+ * ContractAgreementException used for failures in
+ * {@link org.eclipse.tractusx.irs.edc.client.contract.service.EdcContractAgreementService}
  */
-@Getter
-public class UsagePolicyException extends EdcClientException {
+public class ContractAgreementException extends EdcClientException {
 
-    private final transient Policy policy;
-    private final String businessPartnerNumber;
-
-    public UsagePolicyException(final String itemId, final Policy policy, final String businessPartnerNumber) {
-        super("Consumption of asset '" + itemId
-                + "' is not permitted as the required catalog offer policies do not comply with defined IRS policies.");
-        this.policy = policy;
-        this.businessPartnerNumber = businessPartnerNumber;
+    public ContractAgreementException(final String message) {
+        super(message);
     }
 }